feat: added method payment selection

+++ app/src/main/java/com/soleel/paymentapp/MainNavigation.kt
@@ -56,7 +56,7 @@ fun PaymentAppNavigationGraph() {
             )
 
             homeNavigationGraph(
-                navigateToSalesProcessGraph = { calculatorTotal: Int ->
+                navigateToSalesProcessGraph = { calculatorTotal: Float ->
                     navHostController.navigate(SalesProcessGraph(calculatorTotal))
                 }
             )
